resolved merge conflicts. with weapon teams, remade weapon test with projectile services.
