+++ source/core/src/main/com/csse3200/game/components/player/WeaponComponent.java
@@ -311,8 +311,7 @@ public class WeaponComponent extends Component {
                 // Shooting
                 this.setAmmo(-1);
                 // Spawn projectile
-                ProjectileFactory.createProjectile(this.bulletConfig,
-                        this.getEntity().getPosition(), direction);
+                ProjectileFactory.createProjectile(this.bulletConfig, direction);
                 logger.info("Ranged weapon shoot");
             }
             // Reset lastAtttack time
