+++ source/core/src/test/com/csse3200/game/components/player/inventory/WeaponComponentTest.java
@@ -3,12 +3,27 @@ package com.csse3200.game.components.player;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.math.Vector2;
 import com.csse3200.game.components.player.inventory.Collectible;
+import com.csse3200.game.entities.EntityService;
+import com.csse3200.game.entities.configs.ProjectileConfig;
+import com.csse3200.game.entities.factories.ProjectileFactory;
+import com.csse3200.game.physics.PhysicsService;
+import com.csse3200.game.rendering.DebugRenderer;
+import com.csse3200.game.rendering.RenderService;
+import com.csse3200.game.services.GameTime;
+import com.csse3200.game.services.ResourceService;
+import com.csse3200.game.services.ServiceLocator;
+import com.csse3200.game.utils.math.Vector2Utils;
 import org.junit.Test;
 import com.csse3200.game.entities.Entity;
+import org.junit.jupiter.api.BeforeEach;
 
 import static org.junit.Assert.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class WeaponComponentTest {
+
+
     @Test
     public void testWeaponComponent() {
         // create a weapon component
@@ -240,61 +255,6 @@ public class WeaponComponentTest {
         weaponComponent.setReloadTime(Integer.MAX_VALUE);
         // check if weapon component reload time is correct
         assertEquals(Integer.MAX_VALUE, weaponComponent.getReloadTime());
-    }
-
-
-    @Test
-    public void testRangeWeaponReload() {
-        int maxAmmo = 10;
-        // create a weapon component
-        WeaponComponent weaponComponent = new WeaponComponent(new Sprite(),
-                Collectible.Type.RANGED_WEAPON, 10, 5, 1, 0, maxAmmo, 1);
-        // Create test entity to attach weaponcomponent
-        Entity testEntity = new Entity();
-        testEntity.addComponent(weaponComponent);
-        // Ammo is 0
-        assertEquals(0, weaponComponent.getAmmo());
-        // Shot in default direction with ammo at 0
-        weaponComponent.shoot(new Vector2());
-        try {
-            Thread.sleep(800);
-            weaponComponent.shoot(new Vector2());
-            // Weapon is still reloading, weapon should not shoot.
-            assertEquals(maxAmmo, weaponComponent.getAmmo());
-            Thread.sleep(200);
-            weaponComponent.shoot(new Vector2());
-            // Attempt to shoot weapon after reload, weapon should shoot.
-            assertEquals(maxAmmo - 1, weaponComponent.getAmmo());
-        } catch (InterruptedException ex) {
-            // sleep() failed
-            fail();
-        }
-    }
-
-    @Test
-    public void testRangeWeaponFireRate() {
-        int maxAmmo = 10;
-        // create a weapon component
-        WeaponComponent weaponComponent = new WeaponComponent(new Sprite(),
-                Collectible.Type.RANGED_WEAPON, 10, 5, 1, maxAmmo, maxAmmo, 2);
-        // Create test entity to attach weaponcomponent
-        Entity testEntity = new Entity();
-        testEntity.addComponent(weaponComponent);
-        // Shot in default direction with ammo at 0
-        weaponComponent.shoot(new Vector2());
-        try {
-            Thread.sleep(800);
-            weaponComponent.shoot(new Vector2());
-            // Attempt to shoot weapon faster than fire rate, weapon should not shoot.
-            assertEquals(maxAmmo - 1, weaponComponent.getAmmo());
-            Thread.sleep(200);
-            weaponComponent.shoot(new Vector2());
-            // Attempt to shoot weapon equal to fire rate, weapon should shoot.
-            assertEquals(maxAmmo - 2, weaponComponent.getAmmo());
-        } catch (InterruptedException ex) {
-            // sleep() failed
-            fail();
-        }
     }
 
 }
